+++ common/src/main/java/mca/Config.java
@@ -74,6 +74,7 @@ public final class Config implements Serializable {
     public float taxesFactor = 1.5f;
     public int taxSeason = 168000;
     public int marriageChance = 5;
+    public int adventurerAtInnChance = 5;
     public int childrenChance = 5;
     public int bountyHunterInterval = 24000;
     public int bountyHunterThreshold = -5;